backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
httpx==0.24.1
aiohttp==3.11.14
starlette==0.36.3
typing-extensions==4.11.0 